backend/requirements.txt
flask==2.0.1
python-dotenv==0.19.0
flask-cors==3.0.10
requests==2.26.0
google-generativeai==0.3.2
